requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
cryptography
bcrypt
python-dotenv
pytest
requests 
pydantic[email]
python-multipart
